+++ src/Driver/Pdo/AbstractPdoCommand.php
@@ -185,12 +185,10 @@ abstract class AbstractPdoCommand extends AbstractCommand implements PdoCommandI
      *
      * It's a wrapper around {@see PDOStatement::execute()} to support transactions and retry handlers.
      *
-     * @param string|null $rawSql Deprecated. Use `null` value. Will be removed in version 2.0.0.
-     *
      * @throws Exception
      * @throws Throwable
      */
-    protected function internalExecute(string|null $rawSql): void
+    protected function internalExecute(): void
     {
         $attempt = 0;
 
@@ -202,7 +200,7 @@ abstract class AbstractPdoCommand extends AbstractCommand implements PdoCommandI
                     && $this->db->getTransaction() === null
                 ) {
                     $this->db->transaction(
-                        fn () => $this->internalExecute($rawSql),
+                        fn () => $this->internalExecute(),
                         $this->isolationLevel
                     );
                 } else {