+++ src/QueryBuilder/AbstractDMLQueryBuilder.php
@@ -16,7 +16,6 @@ use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Query\QueryInterface;
-use Yiisoft\Db\Schema\ColumnSchemaInterface;
 use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Schema\SchemaInterface;
 
@@ -228,7 +227,7 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
      */
     protected function extractColumnNames(array|Iterator $rows, array $columns): array
     {
-        $columns = $this->getNormalizeColumnNames('', $columns);
+        $columns = $this->getNormalizeColumnNames($columns);
 
         if (!empty($columns)) {
             return $columns;
@@ -333,7 +332,7 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
 
         $names = [];
         $placeholders = [];
-        $columns = $this->normalizeColumnNames('', $columns);
+        $columns = $this->normalizeColumnNames($columns);
         $columnSchemas = $this->schema->getTableSchema($table)?->getColumns() ?? [];
 
         foreach ($columns as $name => $value) {
@@ -367,7 +366,7 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
     protected function prepareUpdateSets(string $table, array $columns, array $params = []): array
     {
         $sets = [];
-        $columns = $this->normalizeColumnNames('', $columns);
+        $columns = $this->normalizeColumnNames($columns);
         $columnSchemas = $this->schema->getTableSchema($table)?->getColumns() ?? [];
 
         foreach ($columns as $name => $value) {
@@ -411,7 +410,7 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
         if ($insertColumns instanceof QueryInterface) {
             [$insertNames] = $this->prepareInsertSelectSubQuery($insertColumns);
         } else {
-            $insertNames = $this->getNormalizeColumnNames('', array_keys($insertColumns));
+            $insertNames = $this->getNormalizeColumnNames(array_keys($insertColumns));
 
             $insertNames = array_map(
                 [$this->quoter, 'quoteColumnName'],
@@ -512,34 +511,19 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
     }
 
     /**
-     * @return mixed The typecast value of the given column.
-     *
-     * @deprecated will be removed in version 2.0.0
-     */
-    protected function getTypecastValue(mixed $value, ColumnSchemaInterface $columnSchema = null): mixed
-    {
-        if ($columnSchema) {
-            return $columnSchema->dbTypecast($value);
-        }
-
-        return $value;
-    }
-
-    /**
      * Normalizes the column names.
      *
-     * @param string $table Not used. Could be empty string. Will be removed in version 2.0.0.
      * @param array $columns The column data (name => value).
      *
      * @return array The normalized column names (name => value).
      *
      * @psalm-return array<string, mixed>
      */
-    protected function normalizeColumnNames(string $table, array $columns): array
+    protected function normalizeColumnNames(array $columns): array
     {
         /** @var string[] $columnNames */
         $columnNames = array_keys($columns);
-        $normalizedNames = $this->getNormalizeColumnNames('', $columnNames);
+        $normalizedNames = $this->getNormalizeColumnNames($columnNames);
 
         return array_combine($normalizedNames, $columns);
     }
@@ -547,12 +531,11 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
     /**
      * Get normalized column names
      *
-     * @param string $table Not used. Could be empty string. Will be removed in version 2.0.0.
      * @param string[] $columns The column names.
      *
      * @return string[] Normalized column names.
      */
-    protected function getNormalizeColumnNames(string $table, array $columns): array
+    protected function getNormalizeColumnNames(array $columns): array
     {
         foreach ($columns as &$name) {
             $name = $this->quoter->ensureColumnName($name);