+++ src/Schema/AbstractTableSchema.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Schema;
 
-use Yiisoft\Db\Exception\NotSupportedException;
-
 use function array_keys;
 
 /**
@@ -151,10 +149,5 @@ abstract class AbstractTableSchema implements TableSchemaInterface
     public function foreignKey(string|int $id, array $to): void
     {
         $this->foreignKeys[$id] = $to;
-    }
-
-    public function compositeForeignKey(int $id, string $from, string $to): void
-    {
-        throw new NotSupportedException(static::class . ' does not support composite FK.');
     }
 }